+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+## [0.2.0] - 2024-06-14
+### Updated
+- deploy bank with 10 production validators
+
+## [0.1.2] - 2024-05-10
+### Updated
+- re-build for dummy production validators
+
 ## [0.1.1] - 2024-04-23
 ### Added
 - verifiable builds